Rename vague info.py -> i_tcp_port_selector.py
